requirements.txt
streamlit
python-docx
markdown
beautifulsoup4